SDK: extract ToolSummarizer helper (oh-tab-g9ys)
